LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-infinity-nan-computed-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: testComputedValueGreaterOrLowerThan

Harness Error (FAIL), message = ReferenceError: Can't find variable: testComputedValueGreaterOrLowerThan

FAIL Property width value 'calc(NaN * 1px)' assert_equals: expected "0px" but got "33554428px"
FAIL Property width value 'calc(NaN * 1%)' assert_equals: expected "0px" but got "33554432px"
